requirements-dev.txt
django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
chromadb==0.4.22
ollama==0.1.7
torch>=2.0.0
sentence-transformers==2.3.1
pypdf==4.0.2
python-docx==1.1.0
openpyxl==3.1.2
unstructured>=0.10.0
cryptography==42.0.2
django-cors-headers==4.3.1
python-decouple==3.8
openai-whisper
ffmpeg-python